Moved all support files to support folder
