+++ cmd/prime-api-client/support/get_move_task_order.go
@@ -1,4 +1,4 @@
-package main
+package support
 
 import (
 	"encoding/json"
@@ -14,29 +14,31 @@ import (
 
 	"github.com/transcom/mymove/cmd/prime-api-client/utils"
 
-	mtoShipment "github.com/transcom/mymove/pkg/gen/supportclient/mto_shipment"
+	mto "github.com/transcom/mymove/pkg/gen/supportclient/move_task_order"
 )
 
-func initUpdateMTOShipmentStatusFlags(flag *pflag.FlagSet) {
+// InitGetMTOFlags declares which flags are enabled
+func InitGetMTOFlags(flag *pflag.FlagSet) {
 	flag.String(utils.FilenameFlag, "", "Name of the file being passed in")
 
 	flag.SortFlags = false
 }
 
-func checkUpdateMTOShipmentStatusConfig(v *viper.Viper, args []string, logger *log.Logger) error {
+func checkGetMTOConfig(v *viper.Viper, args []string, logger *log.Logger) error {
 	err := utils.CheckRootConfig(v)
 	if err != nil {
 		logger.Fatal(err)
 	}
 
 	if v.GetString(utils.FilenameFlag) == "" && (len(args) < 1 || len(args) > 0 && !utils.ContainsDash(args)) {
-		logger.Fatal(errors.New("support-update-mto-shipment-status expects a file to be passed in"))
+		logger.Fatal(errors.New("get-mto expects a file to be passed in"))
 	}
 
 	return nil
 }
 
-func updateMTOShipmentStatus(cmd *cobra.Command, args []string) error {
+// GetMTO creates a gateway and sends the request to the endpoint
+func GetMTO(cmd *cobra.Command, args []string) error {
 	v := viper.New()
 
 	//  Create the logger
@@ -49,31 +51,33 @@ func updateMTOShipmentStatus(cmd *cobra.Command, args []string) error {
 	}
 
 	// Check the config before talking to the CAC
-	err := checkUpdateMTOShipmentStatusConfig(v, args, logger)
+	err := checkGetMTOConfig(v, args, logger)
 	if err != nil {
 		logger.Fatal(err)
 	}
 
 	// Decode json from file that was passed in
 	filename := v.GetString(utils.FilenameFlag)
-	var updateMTOShipmentParams mtoShipment.UpdateMTOShipmentStatusParams
-	err = utils.DecodeJSONFileToPayload(filename, utils.ContainsDash(args), &updateMTOShipmentParams)
+	var getMTOParams mto.GetMoveTaskOrderParams
+	err = utils.DecodeJSONFileToPayload(filename, utils.ContainsDash(args), &getMTOParams)
 	if err != nil {
 		logger.Fatal(err)
 	}
-	updateMTOShipmentParams.SetTimeout(time.Second * 30)
+	getMTOParams.SetTimeout(time.Second * 30)
 
 	// Create the client and open the cacStore
 	supportGateway, cacStore, errCreateClient := utils.CreateSupportClient(v)
 	if errCreateClient != nil {
 		return errCreateClient
 	}
+
 	// Defer closing the store until after the API call has completed
 	if cacStore != nil {
 		defer cacStore.Close()
 	}
+	getMTOParams.SetTimeout(time.Second * 30)
 
-	resp, err := supportGateway.MtoShipment.UpdateMTOShipmentStatus(&updateMTOShipmentParams)
+	resp, err := supportGateway.MoveTaskOrder.GetMoveTaskOrder(&getMTOParams)
 	if err != nil {
 		return utils.HandleGatewayError(err, logger)
 	}
