+++ cmd/prime-api-client/support/update_mto_shipment_status.go
@@ -1,4 +1,4 @@
-package main
+package support
 
 import (
 	"encoding/json"
@@ -9,27 +9,36 @@ import (
 	"time"
 
 	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 
 	"github.com/transcom/mymove/cmd/prime-api-client/utils"
 
-	paymentRequest "github.com/transcom/mymove/pkg/gen/supportclient/payment_requests"
+	mtoShipment "github.com/transcom/mymove/pkg/gen/supportclient/mto_shipment"
 )
 
-func checkListMTOPaymentRequestsConfig(v *viper.Viper, args []string, logger *log.Logger) error {
+// InitUpdateMTOShipmentStatusFlags declares which flags are enabled
+func InitUpdateMTOShipmentStatusFlags(flag *pflag.FlagSet) {
+	flag.String(utils.FilenameFlag, "", "Name of the file being passed in")
+
+	flag.SortFlags = false
+}
+
+func checkUpdateMTOShipmentStatusConfig(v *viper.Viper, args []string, logger *log.Logger) error {
 	err := utils.CheckRootConfig(v)
 	if err != nil {
 		logger.Fatal(err)
 	}
 
 	if v.GetString(utils.FilenameFlag) == "" && (len(args) < 1 || len(args) > 0 && !utils.ContainsDash(args)) {
-		logger.Fatal(errors.New("support-list-mto-payment-requests expects a file to be passed in"))
+		logger.Fatal(errors.New("support-update-mto-shipment-status expects a file to be passed in"))
 	}
 
 	return nil
 }
 
-func listMTOPaymentRequests(cmd *cobra.Command, args []string) error {
+// UpdateMTOShipmentStatus creates a gateway and sends the request to the endpoint
+func UpdateMTOShipmentStatus(cmd *cobra.Command, args []string) error {
 	v := viper.New()
 
 	//  Create the logger
@@ -42,19 +51,19 @@ func listMTOPaymentRequests(cmd *cobra.Command, args []string) error {
 	}
 
 	// Check the config before talking to the CAC
-	err := checkListMTOPaymentRequestsConfig(v, args, logger)
+	err := checkUpdateMTOShipmentStatusConfig(v, args, logger)
 	if err != nil {
 		logger.Fatal(err)
 	}
 
 	// Decode json from file that was passed in
 	filename := v.GetString(utils.FilenameFlag)
-	var paymentReqParams paymentRequest.ListMTOPaymentRequestsParams
-	err = utils.DecodeJSONFileToPayload(filename, utils.ContainsDash(args), &paymentReqParams)
+	var updateMTOShipmentParams mtoShipment.UpdateMTOShipmentStatusParams
+	err = utils.DecodeJSONFileToPayload(filename, utils.ContainsDash(args), &updateMTOShipmentParams)
 	if err != nil {
 		logger.Fatal(err)
 	}
-	paymentReqParams.SetTimeout(time.Second * 30)
+	updateMTOShipmentParams.SetTimeout(time.Second * 30)
 
 	// Create the client and open the cacStore
 	supportGateway, cacStore, errCreateClient := utils.CreateSupportClient(v)
@@ -66,8 +75,7 @@ func listMTOPaymentRequests(cmd *cobra.Command, args []string) error {
 		defer cacStore.Close()
 	}
 
-	// Make the API Call
-	resp, err := supportGateway.PaymentRequests.ListMTOPaymentRequests(&paymentReqParams)
+	resp, err := supportGateway.MtoShipment.UpdateMTOShipmentStatus(&updateMTOShipmentParams)
 	if err != nil {
 		return utils.HandleGatewayError(err, logger)
 	}
