+++ cmd/prime-api-client/support/update_payment_request_status.go
@@ -1,4 +1,4 @@
-package main
+package support
 
 import (
 	"encoding/json"
@@ -14,29 +14,31 @@ import (
 
 	"github.com/transcom/mymove/cmd/prime-api-client/utils"
 
-	mtoserviceitem "github.com/transcom/mymove/pkg/gen/supportclient/mto_service_item"
+	paymentRequest "github.com/transcom/mymove/pkg/gen/supportclient/payment_requests"
 )
 
-func initUpdateMTOServiceItemStatusFlags(flag *pflag.FlagSet) {
+// InitUpdatePaymentRequestStatusFlags declares which flags are enabled
+func InitUpdatePaymentRequestStatusFlags(flag *pflag.FlagSet) {
 	flag.String(utils.FilenameFlag, "", "Name of the file being passed in")
 
 	flag.SortFlags = false
 }
 
-func checkUpdateMTOServiceItemStatusConfig(v *viper.Viper, args []string, logger *log.Logger) error {
+func checkUpdatePaymentRequestStatusConfig(v *viper.Viper, args []string, logger *log.Logger) error {
 	err := utils.CheckRootConfig(v)
 	if err != nil {
 		logger.Fatal(err)
 	}
 
 	if v.GetString(utils.FilenameFlag) == "" && (len(args) < 1 || len(args) > 0 && !utils.ContainsDash(args)) {
-		logger.Fatal(errors.New("support-update-mto-service-item-status expects a file to be passed in"))
+		logger.Fatal(errors.New("update-mto-shipment expects a file to be passed in"))
 	}
 
 	return nil
 }
 
-func updateMTOServiceItemStatus(cmd *cobra.Command, args []string) error {
+// UpdatePaymentRequestStatus creates a gateway and sends the request to the endpoint
+func UpdatePaymentRequestStatus(cmd *cobra.Command, args []string) error {
 	v := viper.New()
 
 	//  Create the logger
@@ -49,19 +51,19 @@ func updateMTOServiceItemStatus(cmd *cobra.Command, args []string) error {
 	}
 
 	// Check the config before talking to the CAC
-	err := checkUpdateMTOServiceItemStatusConfig(v, args, logger)
+	err := checkUpdatePaymentRequestStatusConfig(v, args, logger)
 	if err != nil {
 		logger.Fatal(err)
 	}
 
-	// Decode json from file that was passed into MTO Service item
+	// Decode json from file that was passed in
 	filename := v.GetString(utils.FilenameFlag)
-	var updateServiceItemParams mtoserviceitem.UpdateMTOServiceItemStatusParams
-	err = utils.DecodeJSONFileToPayload(filename, utils.ContainsDash(args), &updateServiceItemParams)
+	var paymentReqParams paymentRequest.UpdatePaymentRequestStatusParams
+	err = utils.DecodeJSONFileToPayload(filename, utils.ContainsDash(args), &paymentReqParams)
 	if err != nil {
 		logger.Fatal(err)
 	}
-	updateServiceItemParams.SetTimeout(time.Second * 30)
+	paymentReqParams.SetTimeout(time.Second * 30)
 
 	// Create the client and open the cacStore
 	supportGateway, cacStore, errCreateClient := utils.CreateSupportClient(v)
@@ -74,12 +76,11 @@ func updateMTOServiceItemStatus(cmd *cobra.Command, args []string) error {
 	}
 
 	// Make the API Call
-	resp, err := supportGateway.MtoServiceItem.UpdateMTOServiceItemStatus(&updateServiceItemParams)
+	resp, err := supportGateway.PaymentRequests.UpdatePaymentRequestStatus(&paymentReqParams)
 	if err != nil {
 		return utils.HandleGatewayError(err, logger)
 	}
 
-	// Get the successful response payload and convert to json for output
 	payload := resp.GetPayload()
 	if payload != nil {
 		payload, errJSONMarshall := json.Marshal(payload)
